refactor: standardize backend middlewares directory
