--- backend/src/middleware/auth.middleware.ts
@@ -1,3 +1,0 @@
-export { authenticateJWT } from '../middlewares/authenticateJWT';
-export { requireRole } from '../middlewares/requireRole';
-export { checkStationAccess } from '../middlewares/checkStationAccess';
