+++ backend/src/middlewares/auth.middleware.ts
@@ -1,0 +1,3 @@
+export { authenticateJWT } from './authenticateJWT';
+export { requireRole } from './requireRole';
+export { checkStationAccess } from './checkStationAccess';
